requirements.txt
requests
python-magic
tensorflow
scikit-learn